[ios] handle weirdness with node-plist
If the plist file has a top-level <plist> element, the contents are read
as an array (with a dict as plist[0]). If it does not, it is read as an
object (the expected behavior).

I'll have to check out if there's a plist spec and what/how this should
be handled. Dumb fix for now