Overhaul dependency uninstallation
This fixes several long-standing bugs with plugman deleting
dependencies. The new rules are to uninstall all of the target plugin's
dependencies, provided they are:
- Not installed at the top level.
- Not depended on by any other plugins.

Similarly, plugin directories are only deleted when no plugin depends on
them, and they are not top-level, on any platform.

Fixes CB-5056, CB-5060, CB-4992, and CB-4189.