CB-5006 Change searchpath to use repeated flag instead of delimiter
The colon can't be used as delimiter for Windows paths like C:\...
Instead of using a delimiter, this change allows to repeat the
--searchpath flag for each additional directory to be searched.
The value for each path is resolved relative to the current dir.
Example:
--searchpath ../my_plugins --searhcpath /usr/lib/other_plugins

The dirs are searched in order. Search is not recursive and will only
look one level down. It will look for
my_plugins/plugin_x/plugin.xml
but not for
my_plugins/bunch_of_plugins/plugin_y/plugin.xml